IOS-4118 Delete file
