+++ Anytype/Sources/Analytics/AnytypeAnalytics/AnytypeAnalytics.swift
@@ -1,67 +1,103 @@
-//import Foundation
-//import Services
-//
-//final class AnytypeAnalytics: Sendable {
-//    
-//    private static let anytypeAnalytics = AnytypeAnalytics()
-//    
-//    static func instance() -> AnytypeAnalytics {
-//        return AnytypeAnalytics.anytypeAnalytics
-//    }
-//    
-//    func setIsEnabled(_ isEnabled: Bool) {
-//        Task {
-//            await AnytypeAnalyticsCore.instance().setIsEnabled(isEnabled)
-//        }
-//    }
-//    
-//    func setEventHandler(_ eventHandler: (@Sendable (_ eventType: String, _ eventProperties: [AnyHashable : Any]?) -> Void)?) {
-//        Task {
-//            await AnytypeAnalyticsCore.instance().setEventHandler(eventHandler)
-//        }
-//    }
-//    
-//    func setEventConfiguartion(event: String, configuation: EventConfigurtion) {
-//        Task {
-//            await AnytypeAnalyticsCore.instance().setEventConfiguartion(event: event, configuation: configuation)
-//        }
-//    }
-//
-//    func initializeApiKey(_ apiKey: String) {
-//        Task {
-//            await AnytypeAnalyticsCore.instance().initializeApiKey(apiKey)
-//        }
-//    }
-//
-//    func setUserId(_ userId: String) async {
-//        await AnytypeAnalyticsCore.instance().setUserId(userId)
-//    }
-//
-//    func setNetworkId(_ networkId: String) async {
-//        await AnytypeAnalyticsCore.instance().setNetworkId(networkId)
-//    }
-//    
-//    func setMembershipTier(tier: MembershipTier?) async {
-//        await AnytypeAnalyticsCore.instance().setMembershipTier(tier: tier)
-//    }
-//    
-//    func logEvent(_ eventType: String, spaceId: String, withEventProperties eventProperties: [AnyHashable : Any]?) {
-//        Task {
-//            await AnytypeAnalyticsCore.instance().logEvent(eventType, spaceId: spaceId, withEventProperties: eventProperties)
-//        }
-//    }
-//    
-//    func logEvent(_ eventType: String, withEventProperties eventProperties: [AnyHashable : Any]?) {
-//        Task {
-//            await AnytypeAnalyticsCore.instance().logEvent(eventType, withEventProperties: eventProperties)
-//        }
-//    }
-//
-//    func logEvent(_ eventType: String) {
-//        logEvent(eventType, withEventProperties: nil)
-//    }
-//    
-//    func logEvent(_ eventType: String, spaceId: String) {
-//        logEvent(eventType, spaceId: spaceId, withEventProperties: nil)
-//    }
-//}
+import Amplitude
+import Services
+
+final class AnytypeAnalytics: @unchecked Sendable {
+
+    private enum Keys {
+        static let interfaceLang = "interfaceLang"
+        static let networkId = "networkId"
+        static let tier = "tier"
+    }
+    
+    private var isEnabled: Bool = true
+    private var eventHandler: ((_ eventType: String, _ eventProperties: [AnyHashable : Any]?) -> Void)?
+    
+    private static let anytypeAnalytics = AnytypeAnalytics()
+
+    private var eventsConfiguration: [String: EventConfigurtion] = [:]
+    private var lastEvents: String = ""
+    private var userProperties: [AnyHashable: Any] = [:]
+    
+    @Injected(\.participantSpacesStorage)
+    private var participantSpacesStorage: any ParticipantSpacesStorageProtocol
+    
+    private init() {
+        // Disable IDFA/IPAddress for Amplitude
+        if let trackingOptions = AMPTrackingOptions().disableIDFA().disableIPAddress() {
+            Amplitude.instance().setTrackingOptions(trackingOptions)
+        }
+        Amplitude.instance().setServerUrl("https://amplitude.anytype.io")
+        userProperties[Keys.interfaceLang] = Locale.current.language.languageCode?.identifier
+    }
+
+    static func instance() -> AnytypeAnalytics {
+        return AnytypeAnalytics.anytypeAnalytics
+    }
+
+    func setIsEnabled(_ isEnabled: Bool) {
+        self.isEnabled = isEnabled
+    }
+    
+    func setEventHandler(_ eventHandler: ((_ eventType: String, _ eventProperties: [AnyHashable : Any]?) -> Void)?) {
+        self.eventHandler = eventHandler
+    }
+    
+    func setEventConfiguartion(event: String, configuation: EventConfigurtion) {
+        eventsConfiguration[event] = configuation
+    }
+
+    func initializeApiKey(_ apiKey: String) {
+        Amplitude.instance().initializeApiKey(apiKey)
+    }
+
+    func setUserId(_ userId: String) {
+        Amplitude.instance().setUserId(userId)
+    }
+
+    func setNetworkId(_ networkId: String) {
+        userProperties[Keys.networkId] = networkId
+    }
+    
+    func setMembershipTier(tier: MembershipTier?) {
+        userProperties[Keys.tier] = tier?.name
+    }
+    
+    func logEvent(_ eventType: String, spaceId: String, withEventProperties eventProperties: @autoclosure () -> [AnyHashable : Any]?) {
+        var eventProperties = eventProperties() ?? [:]
+        let participantSpaceView = participantSpacesStorage.participantSpaceView(spaceId: spaceId)
+        
+        if let permissions = participantSpaceView?.participant?.permission.analyticsType {
+            eventProperties[AnalyticsEventsPropertiesKey.permissions] = permissions.rawValue
+        }
+        
+        if let spaceType = participantSpaceView?.spaceView.spaceAccessType?.analyticsType {
+            eventProperties[AnalyticsEventsPropertiesKey.spaceType] = spaceType.rawValue
+        }
+        
+        logEvent(eventType, withEventProperties: eventProperties)
+    }
+    
+    func logEvent(_ eventType: String, withEventProperties eventProperties: @autoclosure () -> [AnyHashable : Any]?) {
+        let eventProperties = eventProperties()
+        let eventConfiguration = eventsConfiguration[eventType]
+
+        if case .notInRow = eventConfiguration?.threshold, lastEvents == eventType {
+            return
+        }
+
+        lastEvents = eventType
+        
+        eventHandler?(eventType, eventProperties)
+        
+        guard isEnabled else { return }
+        Amplitude.instance().logEvent(eventType, withEventProperties: eventProperties, withUserProperties: userProperties)
+    }
+
+    func logEvent(_ eventType: String) {
+        logEvent(eventType, withEventProperties: nil)
+    }
+    
+    func logEvent(_ eventType: String, spaceId: String) {
+        logEvent(eventType, spaceId: spaceId, withEventProperties: nil)
+    }
+}
